+++ Example/SwiftDemo/SwiftDemo/DetailViewController.swift
@@ -12,6 +12,11 @@ import LoginRadiusSDK
 /* Google Native SignIn
 import GoogleSignIn
 */
+
+/* Twitter Native Sign in
+import TwitterKit
+*/
+
 class DetailViewController: FormViewController {
     
     //List of countries provided from Apple's NSLocale class
@@ -48,7 +53,7 @@ class DetailViewController: FormViewController {
         NotificationCenter.default.removeObserver(self)
     }
     
-    func setupForm(){
+    @objc func setupForm(){
         guard let userAccessToken = LoginRadiusSDK.sharedInstance().session.accessToken
         else
         {
@@ -365,12 +370,27 @@ class DetailViewController: FormViewController {
     
     func logoutPressed()
     {
-        /* Google Native SignIn
+        /* Google Native Sign in
         GIDSignIn.sharedInstance().signOut()
         */
+
+        /* Twitter Native Sign in
+        twitterLogout()
+        */
+        
         LoginRadiusSDK.logout()
         let _ = self.navigationController?.popViewController(animated: true)
     }
+    
+    /* Twitter Native Sign in
+    func twitterLogout(){
+        if let twitterSessions = Twitter.sharedInstance().sessionStore.existingUserSessions() as? [TWTRAuthSession]{
+            for session in twitterSessions{
+                Twitter.sharedInstance().sessionStore.logOutUserID(session.userID)
+            }
+        }
+    }
+    */
     
     func toggleRedBorderShowErrorMessage(cell:UIView, row:BaseRow)
     {